==> SCRIPTS/02_ID_mapping_AR_GSE89408.Rmd <==
---
title: "Mapeo Identificadores"
date: "3/16/2021"
output: html_document
---

RNA-seq

```{r libraries}
library(Biobase)
# BiocManager::install("illuminaHumanv2.db")
#library(illuminaHumanv2.db) # Este es el que hay que cambiar en función de la plataforma
source("medianreps.R")
library(stringr)
```

```{r load_data}
load(file = "GSE_AR_annotated.RData")
GSE89408 <- GSE[["GSE89408"]]
```

# Estandarización de la nomenclatura

```{r estandarizacion}
matriz <- exprs(GSE89408)
muestras <- pData(GSE89408)

muestras <- muestras[,c("Sex:ch1", "disease:ch1", "age:ch1", "acpa positivity:ch1")]
colnames(muestras) <- c("Sex", "Diagnosis", "Age", "Serotype")
muestras$Age <- as.numeric(muestras$Age)
muestras$State <- sapply(muestras$Diagnosis, function(x)
  if (str_detect(x, ".*Normal.*")) {
    "Control"
  } else {
    if (str_detect(x, ".*early.*")) {
      "Early"
    } else {
      "Established"}
})
muestras$Diagnosis <- sapply(muestras$Diagnosis, function(x) ifelse(str_detect(x, ".*Normal.*"), "Control", "RA"))
muestras$Diagnosis <- as.factor(muestras$Diagnosis)
muestras$Sex <- as.factor(gsub('F', 'Female', muestras$Sex))
muestras$Sex <- as.factor(gsub('M', 'Male', muestras$Sex))
```

```{r diagnosis_sex}
muestras$Diagnosis_Sex <- factor(paste(muestras$Diagnosis, muestras$Sex, sep = "_"))
```

# Anotación

RNA-seq

Biomart:

```{r}
library(biomaRt)
ensembl <-  useMart("ENSEMBL_MART_ENSEMBL")
    # listDatasets(ensembl) # function to see which datasets are present in ensembl
ensembl <- useDataset("hsapiens_gene_ensembl", mart = ensembl)
    # listFilters(ensembl) # check which filters are available
    # listAttributes(ensembl) # check attributes are available to select
annot <- getBM(attributes = c("ensembl_gene_id", "hgnc_symbol", "description", "entrezgene_id"),
                 filters = "hgnc_symbol",
                 values = rownames(matriz),
                 mart = ensembl)
indices <- BiocGenerics::match(rownames(matriz), annot$hgnc_symbol)
fData(GSE89408) <- annot[indices,]
rownames(matriz) <- fData(GSE89408)$entrezgene_id
matriz <- matriz[!is.na(rownames(matriz)),]
matriz <- medianReps(matriz)
indices <- BiocGenerics::match(rownames(matriz), fData(GSE89408)$entrezgene_id)
ENTREZIDS <- fData(GSE89408)$entrezgene_id[indices]
featuredata <- fData(GSE89408)[indices,]
```

Vamos a sustituir la información original en el fichero GSE por la información actualizada y guardarlo en un nuevo fichero.

```{r}
# GSE update
experimentData <- GSE$GSE89408@experimentData
assayData <- matriz
phenoData <- new("AnnotatedDataFrame", data = muestras, varMetadata = data.frame(labelDescription = rep(NA, ncol(muestras)), row.names = colnames(muestras)))
annotation <- GSE$GSE89408@annotation
featureData <- featuredata
rownames(featuredata) <- featuredata$entrezgene_id
GSE89408n <- new("ExpressionSet", exprs= assayData, phenoData = phenoData, experimentData = experimentData, annotation = annotation, featureData = as(featuredata, "AnnotatedDataFrame"))
GSE$GSE89408 <- GSE89408n

# Guardar
save(GSE, file = "GSE_AR_annotated.RData")
rm(assayData, experimentData, GSE89408, phenoData, annotation)
```
